+++ Constants.hpp
@@ -18,6 +18,7 @@
 
 // ESP32 based boards
 #define BOARD_ESP32_ESP32DEV 1001
+#define BOARD_OAE_V1         1002
 
 /**
  * Supported keypad/display types. Use one of these values for DISPLAY_TYPE configuration matching your used display and keypad.