+++ Changelog.md
@@ -1,4 +1,8 @@
+**V1.13.18 - Updates**
+- Final changes for OAE
+
 **V1.13.17 - Updates**
+- Store AZ/ALT steps per degree and allow them to be set from Meade
 - Added more output to the InfoDisplay during boot and fixed a bug in console mode.
 - Fixed a bug that was causing the firmware to hang after a slew, if a :Q# command was issued (thanks to user c3n for tracking it down, pun intended).
 
@@ -7,6 +11,7 @@
 - Guide pulses are now ignored for DEC as well when at the limits.
 
 **V1.13.15 - Updates**
+- Changes necessary for OAE, add OAE board
 - Check `INFO_DISPLAY_TYPE` builds in CI
 - Fix `INFO_DISPLAY_TYPE_I2C_SSD1306_128x64` for esp32 builds